requirements.txt
aiohttp
asyncio
beautifulsoup4
lxml
psycopg2-binary
python-dotenv
APScheduler
